+++ example/peripheral/i2s/sdkconfig.h
@@ -28,11 +28,11 @@
 #define CONFIG_GCC_CODE_MODEL_SMALL
 /* CONFIG_GCC_CODE_MODEL_LARGE is not set */
 /* end of Compiler configuration */
-#define CONFIG_USE_CACHE
-#define CONFIG_USE_MMU
 #define CONFIG_BOOT_WITH_FLUSH_CACHE
 /* CONFIG_MMU_DEBUG_PRINTS is not set */
 /* end of Arm architecture configuration */
+#define CONFIG_MMU_PAGE_SIZE 0x1000
+#define CONFIG_MAX_XLAT_TABLES 256
 /* end of Arch configuration */
 
 /* Soc configuration */
@@ -43,6 +43,7 @@
 /* CONFIG_TARGET_E2000S is not set */
 /* CONFIG_TARGET_FT2004 is not set */
 /* CONFIG_TARGET_D2000 is not set */
+/* CONFIG_TARGET_PD2308 is not set */
 #define CONFIG_SOC_NAME "e2000"
 #define CONFIG_TARGET_TYPE_NAME "q"
 #define CONFIG_SOC_CORE_NUM 4
@@ -51,7 +52,6 @@
 #define CONFIG_F64BIT_MEMORY_ADDRESS 0x2000000000
 #define CONFIG_F64BIT_MEMORY_LENGTH 0x800000000
 #define CONFIG_TARGET_E2000
-/* CONFIG_USE_SPINLOCK is not set */
 #define CONFIG_DEFAULT_DEBUG_PRINT_UART1
 /* CONFIG_DEFAULT_DEBUG_PRINT_UART0 is not set */
 /* CONFIG_DEFAULT_DEBUG_PRINT_UART2 is not set */
@@ -84,6 +84,7 @@
 
 /* Sdk common configuration */
 
+#define CONFIG_ELOG_LINE_BUF_SIZE 0x100
 /* CONFIG_LOG_VERBOS is not set */
 /* CONFIG_LOG_DEBUG is not set */
 /* CONFIG_LOG_INFO is not set */
@@ -150,6 +151,7 @@
 #define CONFIG_USE_ES8336
 #define CONFIG_USE_FI2S
 /* end of I2S Configuration */
+/* CONFIG_USE_I3C is not set */
 /* end of Drivers configuration */
 
 /* Build setup */
@@ -228,11 +230,6 @@
 /* CONFIG_FREERTOS_USE_XMAC is not set */
 /* CONFIG_FREERTOS_USE_GMAC is not set */
 /* end of Freertos Eth Drivers */
-
-/* Freertos Gpio Drivers */
-
-/* CONFIG_FREERTOS_USE_GPIO is not set */
-/* end of Freertos Gpio Drivers */
 
 /* Freertos Spim Drivers */